+++ config/breadcrumbs/backoffice.rb
@@ -29,6 +29,11 @@ crumb :backoffice_service_edit do |service|
   parent :backoffice_service, service
 end
 
+crumb :backoffice_service_preview do |service|
+  link "Preview", preview_backoffice_service_path(service)
+  parent :backoffice_service_edit, service
+end
+
 crumb :backoffice_research_areas do
   link "Research Areas", backoffice_research_areas_path
   parent :backoffice_root